// types/Funding.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@0.19.0.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export interface InstantiateMsg {
  blacklist_contract_address: string;
  cw20_donation_token_address?: string | null;
  general_fund_address: string;
  native_donation_token_denom: string;
  spark_token_contract_address: string;
}
export type ExecuteMsg = {
  add_campaign: {
    campaign_address: string;
    campaign_name: string;
  };
} | {
  disable_campaign: {
    campaign_name: string;
  };
} | {
  receive: Cw20ReceiveMsg;
} | {
  fund: FundMsg;
} | {
  admin_fund: {
    amount: Uint128;
    fund_msg: FundMsg;
    to: string;
  };
} | {
  update_admins: {
    admins: Addr[];
  };
} | {
  update_config: {
    blacklist_contract_address?: string | null;
    cw20_donation_token_address?: string | null;
    general_fund_address?: string | null;
    native_token_denom?: string | null;
    spark_token_contract_address?: string | null;
  };
} | {
  update_nickname: {
    nickname?: string | null;
  };
} | {
  update_validator_link: {
    validator_link?: Link | null;
  };
};
export type Uint128 = string;
export type Binary = string;
export type FundMsg = {
  fund_general: {
    donor_address_type: AddressType;
    on_behalf_of?: string | null;
  };
} | {
  fund_campaign: {
    campaign_name: string;
    donor_address_type: AddressType;
    on_behalf_of?: string | null;
  };
};
export type AddressType = "private" | "validator";
export type Addr = string;
export interface Cw20ReceiveMsg {
  amount: Uint128;
  msg: Binary;
  sender: string;
}
export interface Link {
  label: string;
  url: string;
}
export type QueryMsg = {
  query_total_donated: {};
} | {
  query_get_campaign: {
    campaign_name: string;
  };
} | {
  query_get_donor: {
    donor_address: string;
  };
};
export type Status = "active" | "inactive";
export interface GetCampaignResponse {
  campaign: Campaign;
}
export interface Campaign {
  address: Addr;
  id: number;
  name: string;
  status: Status;
  total_donations: Uint128;
}
export interface GetDonorResponse {
  donor: Donor;
}
export interface Donor {
  addr: Addr;
  nickname?: string | null;
  total_campaign_donations: Uint128;
  total_general_donations: Uint128;
  validator_link?: Link | null;
}